refactor: separate paymentMethod input item
