+++ components/InputBar/index.js
@@ -1,5 +1,6 @@
 import createAmount from './Items/amount.js';
 import createContent from './Items/content.js';
+import createPaymentMethod from './Items/paymentMethod.js';
 
 function createInputBar() {
     const container = document.createElement('div');
@@ -23,26 +24,7 @@ function createInputBar() {
     const contentItem = createContent();
     container.appendChild(contentItem);
 
-    const paymentMethodItem = document.createElement('div');
-    paymentMethodItem.className = 'input-bar-item';
-    paymentMethodItem.innerHTML = `
-        <div class="input-bar-item-wrapper">
-            <label for="payment-method" class="light-12">
-                결제수단
-            </label>
-        </div>
-        <div class="input-bar-item-wrapper">
-            <div class="select-container semibold-12">
-                선택하세요
-            </div>
-            <img
-                src="/assets/icons/chevron-down.svg"
-                alt="Chevron Down Icon"
-                width="16"
-                height="16"
-            />
-        </div>
-    `;
+    const paymentMethodItem = createPaymentMethod();
     container.appendChild(paymentMethodItem);
 
     const categoryItem = document.createElement('div');
